Movido CLI a un archivo aparte
